refactor(web): 移动消息输入框相关组件到独立文件夹
